components > homepage > LEFT & RIGHT restructured
<components > RIGHT > Homepage_Selector:
is rendering list items for Education / Projects / Education / Resume
