+++ client/src/components/homepage/RIGHT/Homepage_Selector.js
@@ -16,13 +16,32 @@ const Styled_Homepage_Selector = styled.div`
     padding: 10px;
 
     border: 1px dashed black;
+
+    & a {
+        color: black;
+    }
     `
+
+// STATE
+const links = [
+    'Experience', 
+    'Projects',
+    'Education',
+    'Resume',
+]
 
 // __MAIN FUNCTIONAL COMPONENT__
 function Homepage_Selector() {
     return (
         <Styled_Homepage_Selector className="Homepage Selector">
-            Homepage Selector
+            <ul>
+                {links.map((title, key) => {
+                    console.log(title)
+                    return (
+                        <li key={key}><a href="#">{title}</a></li>
+                    )
+                })}
+            </ul>
         </Styled_Homepage_Selector>
     )
 }
