+++ client/src/styles/index.css
@@ -61,3 +61,7 @@ html {
 body {
 	font-size: 62.5%;
 }
+
+a {
+	text-decoration: none;
+}
